+++ .gitignore
@@ -7,3 +7,5 @@
 pubspec.lock
 
 build/
+
+!bin/
