+++ README.md
@@ -1,14 +1,9 @@
 # Design System Generator
 
-This generator (or builder, since it runs with the dart build runner)
-allows you to generate a design system for your project.
+This generator allows you to generate a design system for your project.
 You can use it to generate design tokens based on a JSON configuration.
 
 The configuration can be viewed at [Design System JSON Schema](./design-system.schema.json).
-
-The current version is pretty static (file-wise). The design system json **must**
-be under `lib/design-system.json`, otherwise nothing gets generated. Furthermore,
-the results are all created in `lib/ui/identity/*`.
 
 The generator creates the following elements:
 
@@ -23,8 +18,10 @@ Refer to the [example](./example/example.md) to see how to use the generator.
 
 Summary:
 
-- Create a `design-system.json` file in your `lib` folder
+- Create a `design-system.json` file in your app folder.
+  The file can be passed as config or will be searched in your
+  folder by a glob pattern.
 - Configure your design system to your wishes
-  (add `"$schema": "https://github.com/smartive/flutter-design-system-generator/blob/main/design-system.schema.json"`) to the start to get
-  JSON schema validation
-- Run `flutter pub run build_runner build` to generate the design system
+  (add `"$schema": "https://github.com/smartive/flutter-design-system-generator/blob/main/design-system.schema.json"`)
+  to the start to get JSON schema validation
+- Run `flutter pub run design_system_generator generate` to generate the design system
