--- build.yaml
@@ -1,44 +1,0 @@
-builders:
-  colorBuilder:
-    import: 'package:design_system_generator/builder.dart'
-    builder_factories: ['colorBuilder']
-    build_extensions: { 'design-system.json': ['ui/identity/colors.g.dart'] }
-    build_to: source
-    auto_apply: root_package
-    defaults:
-      generate_for:
-        include:
-          - lib/design-system.json
-
-  spatialBuilder:
-    import: 'package:design_system_generator/builder.dart'
-    builder_factories: ['spatialBuilder']
-    build_extensions: { 'design-system.json': ['ui/identity/spatials.g.dart'] }
-    build_to: source
-    auto_apply: root_package
-    defaults:
-      generate_for:
-        include:
-          - lib/design-system.json
-
-  screenBuilder:
-    import: 'package:design_system_generator/builder.dart'
-    builder_factories: ['screenBuilder']
-    build_extensions: { 'design-system.json': ['ui/identity/screens.g.dart'] }
-    build_to: source
-    auto_apply: root_package
-    defaults:
-      generate_for:
-        include:
-          - lib/design-system.json
-
-  borderRadiusBuilder:
-    import: 'package:design_system_generator/builder.dart'
-    builder_factories: ['borderRadiusBuilder']
-    build_extensions: { 'design-system.json': ['ui/identity/border_radii.g.dart'] }
-    build_to: source
-    auto_apply: root_package
-    defaults:
-      generate_for:
-        include:
-          - lib/design-system.json
