+++ example/example.md
@@ -1,16 +1,15 @@
 # Using the design system generator
 
-To use the design system generator, you need to install it as a dev
-dependency along with the `build_runner` package.
+To use the design system generator, you need to install it as a dev dependency.
 
-`flutter pub install -d build_runner design_system_generator`.
+`flutter pub install -d design_system_generator`.
 
-Then, you need to create a `design-system.json` file in your `lib` folder.
+Then, you need to create a `design-system.json` file in your root folder.
 There is a json schema to enable autocompletion and schema validation for
 your configuration file.
 
 After you have created your configuration file, you can run the generator
-with `flutter pub run build_runner build --delete-conflicting-outputs`.
+with `flutter pub run design_system_generator generate`.
 
 Note that colors are in the format `#RRGGBB` or `#RRGGBBAA` (with alpha).
