--- lib/builder.dart
@@ -1,13 +1,0 @@
-library builder;
-
-import 'package:build/build.dart';
-
-import 'src/border_radius_builder.dart';
-import 'src/color_builder.dart';
-import 'src/screen_builder.dart';
-import 'src/spatial_builder.dart';
-
-Builder colorBuilder(BuilderOptions options) => ColorBuilder();
-Builder spatialBuilder(BuilderOptions options) => SpatialBuilder();
-Builder screenBuilder(BuilderOptions options) => ScreenBuilder();
-Builder borderRadiusBuilder(BuilderOptions options) => BorderRadiusBuilder();
