+++ pubspec.yaml
@@ -13,7 +13,9 @@ environment:
   sdk: '>=2.19.4 <3.0.0'
 
 dependencies:
-  build: ^2.3.1
+  args:
+  path:
+  glob:
 
 dev_dependencies:
   flutter:
